refactor(web): split workspace repository chips
